implemtaion du reste des apis

+++ backend-spring/src/main/java/fr/ubo/kanban/repositories/ColonneRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ColonneRepository extends JpaRepository<Colonne, Integer> {
-    List<Colonne> findByTableauId(Integer tableauId);
+public interface ColonneRepository extends JpaRepository<Colonne, Long> {
+    List<Colonne> findByTableauId(Long tableauId);
 }
